Make github sync a component
